% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_tibble_to_RSE.R
\name{convert_tibble_to_RSE}
\alias{convert_tibble_to_RSE}
\title{convert_tibble_to_RSE}
\usage{
convert_tibble_to_RSE(NomeMatrix, ROIs_gr, annots)
}
\arguments{
\item{NomeMatrix}{A tibble where each row corresponds to a sample ROI combination,
which should contain the following columns:
nFragsAnalyzed = the number fragments (read pairs) that was available
names = the names of the ROIs that were analyzed
SampleName = the names of the samples that were analzyed
GCH_DataMatrix = for every combination of ROI and sample, a matrix where the columns
correspond to genomic positions from start to end of the ROI and the rows to fragments.
1 = protected from GpC methylation. 0 = not protected from GpC methylation. For example, the ouput of the fetch-NOMe R package.}

\item{ROIs_gr}{A GRanges object describing the regions of interest (ROIs).}

\item{annots}{A data.fame with a column for sample names and a column for group names.}
}
\value{
A Ranged Summarized Experiment with an entry for each ROI. The rowData contains information about each ROI,
including a ROIgroup.The assays contain:nFragsAnalyzed, describing the number of fragments that were analyzed for each sample/ROI combination.
reads, containg a Gpos object for each sample/ROI combination, with a position for each base in the ROI and two metadata columns: protection,
a sparse matrix where TRUE stands for Cs protected from methylation, and methylation, where TRUE stands for methylated Cs.
}
\description{
Converts NOMe-seq data in tibble format to Ranged Summarized Experiment format.
}
\details{
Converts NOMe-seq data in tibble format to Ranged Summarized Experiment format.
}
